# docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  controller.CommonResult:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
...
  message.DataExportReq:
    properties:
      accessKey:
        type: string
      bucketUrl:
        type: string
      clusterId:
        type: string
      comment:
        type: string
      endpointUrl:
        type: string
      fileType:
        type: string
      filter:
        type: string
      password:
        type: string
      secretAccessKey:
        type: string
      sql:
        type: string
      storageType:
        type: string
      userName:
        type: string
      zipName:
        type: string
    type: object
  message.DataExportResp:
    properties:
      recordId:
        type: string
      workFlowId:
        description: Asynchronous task workflow ID
        type: string
    type: object
  message.DataImportReq:
    properties:
      accessKey:
        type: string
      bucketUrl:
        type: string
      clusterId:
        type: string
      comment:
        type: string
      endpointUrl:
        type: string
      password:
        type: string
      recordId:
        type: string
      secretAccessKey:
        type: string
      storageType:
        type: string
      userName:
        type: string
    type: object
  message.DataImportResp:
    properties:
      recordId:
        type: string
      workFlowId:
        description: Asynchronous task workflow ID
        type: string
    type: object
...
  message.DeleteImportExportRecordReq:
    type: object
  message.DeleteImportExportRecordResp:
    properties:
      recordId:
        type: string
    type: object
...
  message.QueryDataImportExportRecordsResp:
    properties:
      transportRecords:
        items:
          $ref: '#/definitions/structs.DataImportExportRecordInfo'
        type: array
    type: object
...
  structs.DataImportExportRecordInfo:
    properties:
      clusterId:
        type: string
      comment:
        type: string
      createTime:
        type: string
      deleteTime:
        type: string
      endTime:
        type: string
      filePath:
        type: string
      recordId:
        type: string
      startTime:
        type: string
      status:
        type: string
      storageType:
        type: string
      transportType:
        type: string
      updateTime:
        type: string
      zipName:
        type: string
    type: object
...
host: localhost:4116
info:
  contact:
    email: zhangpeijin@pingcap.com
    name: zhangpeijin
  description: EM UI API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: EM UI API
  version: "1.0"
paths:
...
  /clusters/export:
    post:
      consumes:
      - application/json
      description: export
      parameters:
      - description: cluster info for data export
        in: body
        name: dataExport
        required: true
        schema:
          $ref: '#/definitions/message.DataExportReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.DataExportResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: export data from tidb cluster
      tags:
      - cluster export
  /clusters/import:
    post:
      consumes:
      - application/json
      description: import
      parameters:
      - description: cluster info for import data
        in: body
        name: dataImport
        required: true
        schema:
          $ref: '#/definitions/message.DataImportReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.DataImportResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: import data to tidb cluster
      tags:
      - cluster import
...
  /clusters/transport:
    get:
      consumes:
      - application/json
      description: query records of import and export
      parameters:
      - in: query
        name: clusterId
        type: string
      - in: query
        name: endTime
        type: integer
      - description: Current page location
        in: query
        name: page
        type: integer
      - description: Number of this request
        in: query
        name: pageSize
        type: integer
      - in: query
        name: reImport
        type: boolean
      - in: query
        name: recordId
        type: string
      - in: query
        name: startTime
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.QueryDataImportExportRecordsResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: query records of import and export
      tags:
      - cluster data transport
  /clusters/transport/{recordId}:
    delete:
      consumes:
      - application/json
      description: delete data transport record
      parameters:
      - description: data transport recordId
        in: path
        name: recordId
        required: true
        type: integer
      - description: data transport record delete request
        in: body
        name: DataTransportDeleteReq
        required: true
        schema:
          $ref: '#/definitions/message.DeleteImportExportRecordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.CommonResult'
            - properties:
                data:
                  $ref: '#/definitions/message.DeleteImportExportRecordResp'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controller.CommonResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonResult'
      security:
      - ApiKeyAuth: []
      summary: delete data transport record
      tags:
      - cluster data transport
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"